web/registry/registry/pyproject.toml
[tool.poetry]
name = "vo-registry"
version = "0.3.0dev"
description = "A VORegistry"
authors = ["Destry Saul <destry.saul@noirlab.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
lxml = "^5.0.0"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
xmltodict = "^0.13.0"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.0.0"
requests = "^2.31.0"
flake8 = "^7.0.0"
fastapi = {extras = ["all"], version = "^0.111.0"}
pytest-cov = "^5.0.0"
mypy = "^1.8.0"
lxml-stubs = "^0.5.1"
types-xmltodict = "^0.13.0.3"
types-pyyaml = "^6.0.12.12"
pre-commit = "^3.8.0"

[tool.coverage.run]
branch = true

[tool.coverage.report]
ignore_errors = true
show_missing = true
skip_empty = true
precision = 2

[tool.coverage.xml]
output = "coverage.xml"

[tool.pytest.ini_options]
addopts = "--cov --cov-report term --cov-report xml --junit-xml=report.xml"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
